refac(submit): move congrats modal logic to submitStage
