+++ src/wizards/tokenSwapDealWizard/components/stageButtons/stageButtons.html
@@ -18,7 +18,7 @@
     <pbutton
       if.bind="showSubmit"
       type="primary"
-      click.delegate="wizardService.submit(wizardManager)"
+      click.delegate="onSubmit()"
       data-test="wizard-submit-button">
       Submit
     </pbutton>
