+++ src/wizards/tokenSwapDealWizard/stages/submitStage/submitStage.html
@@ -2,7 +2,7 @@
   <require from="../../components/stageWrapper/stageWrapper"></require>
   <require from="./daoInformation/daoInformation"></require>
 
-  <stage-wrapper header="Submit" wizard-manager.bind="wizardManager">
+  <stage-wrapper header="Submit" wizard-manager.bind="wizardManager" on-submit.call="onSubmit()">
     <div class="stageSection">
       <!-- Sidebar -->
       <div class="stageSectionSidebar">
