+++ src/wizards/tokenSwapDealWizard/stages/submitStage/submitStage.ts
@@ -1,5 +1,6 @@
 import { autoinject } from "aurelia-framework";
 import { IDealRegistrationTokenSwap } from "entities/DealRegistrationTokenSwap";
+import { AlertService, IAlertModel } from "services/AlertService";
 import { IWizardState, WizardService } from "wizards/services/WizardService";
 import { IStageMeta } from "wizards/tokenSwapDealWizard/dealWizardTypes";
 import { WizardManager } from "wizards/tokenSwapDealWizard/wizardManager";
@@ -11,12 +12,28 @@ export class SubmitStage {
   public wizardState: IWizardState<IDealRegistrationTokenSwap>;
   private submitData: IDealRegistrationTokenSwap;
 
-  constructor(private wizardService: WizardService) {}
+  constructor(
+    private wizardService: WizardService,
+    private alertService: AlertService,
+  ) {}
 
   activate(stageMeta: IStageMeta): void {
     this.wizardManager = stageMeta.wizardManager;
     this.wizardState = this.wizardService.getWizardState(this.wizardManager);
 
     this.submitData = this.wizardState.registrationData;
+  }
+
+  public async onSubmit(): Promise<void> {
+    const congratulatePopupModel: IAlertModel = {
+      header: "Your deal has been submitted!",
+      message:
+        "<p class='excitement'>Share your new deal proposal with your community!</p><p class='tweetlink'><a href='https://twitter.com' target='_blank' rel='noopener noreferrer'>TWEET <i class='fab fa-twitter'></i></a></p>",
+      confetti: true,
+      buttonTextPrimary: "Go to deal (todo)",
+      className: "congratulatePopup",
+    };
+
+    await this.alertService.showAlert(congratulatePopupModel);
   }
 }
